pyproject.toml
[tool.poetry]
name = "poirot"
version = "1.0.1"
description = "web spider tools"
authors = ["harumonia <zxjlm233@163.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1.2"
numpy = "1.18.0"
flask-socketio = "^5.0.1"
loguru = "^0.5.3"
opencv-python = "^4.5.1"
onnxruntime = "1.5.2"
shapely = "^1.7.1"
fonttools = "^4.18.2"
requests = "^2.25.1"
pyclipper = "^1.2.1"
pillow = "^8.1.0"
reportlab = "^3.5.59"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "douban"
default = true
url = "https://pypi.doubanio.com/simple/"